src/ingestion/requirements.txt
azure-functions
azure-identity
azure-storage-blob
azure-search-documents
openai
